+++ libs/base_core/src/media/audio/mod.rs
@@ -1,0 +1,22 @@
+// devela_base_core::media::audio
+//
+//! Audio functionality.
+//
+// safety
+// #![cfg_attr(base_safe_audio, forbid(unsafe_code))] // no feature
+
+// mod codec; // …
+mod layout; // AudioChannel, AudioChannels
+// mod pcm; // PcmBuffer, PcmPlanar, PcmRaw, PcmWav, PcmSample, PcmSpec, …
+// mod spatial; // …
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            // codec::*,
+            layout::*,
+            // pcm::*,
+            // spatial::*,
+        };
+    }
+}