+++ src/media/audio/mod.rs
@@ -5,12 +5,15 @@
 // safety
 #![cfg_attr(feature = "safe_audio", forbid(unsafe_code))]
 
-// WIPZONE
 // mod drum_machine;
 
 crate::structural_mods! { // _mods
     _mods {
-        // WIPZONE
         // pub use super::drum_machine::*;
+
+        // re-exports
+        pub use devela_base_core::media::audio::{
+            AudioChannel, AudioChannels,
+        };
     }
 }